[Log] Moving Logger classes to a new Logger namespace for better organization.
