+++ src/Valkyrja/Broadcast/Provider/ServiceProvider.php
@@ -25,7 +25,7 @@ use Valkyrja\Broadcast\PusherBroadcaster;
 use Valkyrja\Container\Contract\Container;
 use Valkyrja\Container\Support\Provider;
 use Valkyrja\Crypt\Contract\Crypt;
-use Valkyrja\Log\Contract\Logger;
+use Valkyrja\Log\Logger\Contract\Logger;
 
 use const CURL_IPRESOLVE_V4;
 use const CURLOPT_IPRESOLVE;
