+++ src/Valkyrja/Log/Logger/NullLogger.php
@@ -11,12 +11,12 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace Valkyrja\Log;
+namespace Valkyrja\Log\Logger;
 
 use Override;
 use Throwable;
-use Valkyrja\Log\Contract\Logger as Contract;
 use Valkyrja\Log\Enum\LogLevel;
+use Valkyrja\Log\Logger\Contract\Logger as Contract;
 
 /**
  * Class NullLogger.
