+++ tests/Functional/ApplicationTest.php
@@ -20,7 +20,7 @@ use Valkyrja\Http\Client\Contract\Client;
 use Valkyrja\Http\Message\Factory\Contract\ResponseFactory;
 use Valkyrja\Http\Routing\Contract\Router;
 use Valkyrja\Http\Server\Contract\RequestHandler;
-use Valkyrja\Log\Contract\Logger;
+use Valkyrja\Log\Logger\Contract\Logger;
 use Valkyrja\Session\Manager\Contract\Session;
 use Valkyrja\Tests\EnvClass;
 use Valkyrja\View\Renderer\Contract\Renderer;
